+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `coverage` command
+  - It outputs the coverage after running `batler test`. For more information, run `batler coverage --help`.
 
 ## [0.2.0] - 2020-12-04
 ### Added